requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor-5==0.2.18
django-jazzmin==3.0.1
django-modeltranslation==0.19.17
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2